[ML] Wait for tasks between ML security REST tests (#1056)
This is basically #963 but for the security tests rather than the
standard integration tests.

There is a slight difference in the code, as the security version
needs to set the correct authorization header for the admin user.

Additionally, a change has been made to the MlRestTestStateCleaner
so that it leaves all index creation to the base class now.  This means
we can wait for pending indexing operations between closing jobs
and deleting indices.