[DOCS] Re-enable XPack doc testing (#1146)
Adds a gradle build for documentation testing, partially stolen from the xpack meta repo. Updated to make work with the current setup

index.asciidoc was updated/enabled as a test, and various pre-existing tests appear to be working. There are a large number of tests still ignored due to missing proper // CONSOLE tags