Reenable OldMonitoringIndicesBackwardsCompatibilityTests on 5.x (#1331)
This test is been ignored for too long because of the way exporters
are designed (see #257). Modifying exporters is a long process and
might potentially break things, therefore a trick has been implemented
in LocalExporterTests few weeks ago to ensure a better shutdown of the
exporters in our tests. This trick seems to work but adds an extra cost
of 5-10 seconds per test: it first stops the monitoring service, then
looks for some monitoring documents and checks the last time they were
indexed, exiting when the last cluster state doc is 2-3 collection
interval old. It indicates that in flight bulk requests have been fully
processed. This is far from ideal (OldMonitoringIndicesBackwardsCompatibilityTests
takes 7 min on my CI) but at least the test run until a better solution
is implemented for exporters (if any).