Fix Authentication BWC when SSL is enabled (#1356)
This commit takes the fix that @tvernum proposed to allow 5.4+ nodes to communicate with 5.3.x and
older nodes when SSL is enabled. In addition, a new QA test is added that uses SSL for transport as
a way to validate rolling upgrades work with ssl. Without the fix, this test fails.