Fix wire compatibility of authentication with system key (#1557)
5.4.0 and 5.4.1 disagree on when a header should be signed and
this was causing 5.4.0 and the tip of the 5.4 branch not to be
able to handshake. The fix is to mimick the 5.4.0 logic when
dealing with 5.4.0. The logic isn't as good, but we need to match
or else the versions can't talk.

The other change here is to rework the rolling restart tests
to use the system key so they would have caught this error.

There are two major flaws with this approach that will have to
be handled in a followup for lack of time:
1. We either need more coverage for rolling restarts with x-pack
or fewer options. We can enable and disable ssl and we can
configure the system key or not. Either we should have 4 sets of
rolling restart tests or we should remove some of these options.
2. The way in which we know if we're "dealing with 5.4.0" is a little
suspect because it uses a variable that is frequently hardcoded
to `Version.CURRENT`. We should come up with some way for headers
to serialize themselves with access to the version of the
serialization channel. Or something.