Rewrite the authentication context when we respond to a 5.4.0 node (#1580)
5.4.0 has a bug that it somehow rejects it's own headers due to a
mix of versions in the authentication headers. This commit special
cases the 5.4.0 version to ensure we can safely upgrade with rolling
restarts form 5.4.0 to newer versions.

This commit also adds a dummy index that has 0 replicas to ensure
we can upgrade indices without HA/replicas. This would have surfaced
the bug fixed in this issue immediately.

Closes #1576