Test: force TLSv1.2 in Netty3HandshakeWaitingHandlerTests
This commit sets the supported protocols to only allow TLSv1.2 as an attempted workaround to the
test failures on windows related to a bad record mac. The forcing of TLSv1.2 means that the RSA
pre master secret fix is enabled by default; whereas in the default configuration the fix is not
enabled and we see the bad record mac error on windows. Additionally, a test issue was noticed
in the order that the write and handshake operations were being called.

Relates #1761